refactor: package 구조 세분화
