+++ src/test/java/laddergame/domain/result/LadderLineResultTest.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.domain.result;
 
 import laddergame.exception.LadderResultIndexNotFoundException;
 import org.junit.jupiter.api.BeforeEach;
